+++ extension/src/pages/MainPage.tsx
@@ -1,87 +1,6 @@
-import { css } from "@emotion/react";
-import styled from "@emotion/styled";
 import React, { ChangeEvent, useEffect, useRef, useState } from "react";
-import { Button, Icon } from "../components";
-
-const Container = styled.div`
-  display: flex;
-  flex-direction: column;
-  width: 355px;
-  height: 380px;
-  padding: 20px;
-  border-radius: 15px;
-  background-color: #ffffff;
-  box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
-`;
-
-const IconWraaper = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-`;
-
-const ImageWrapper = styled.div`
-  align-self: center;
-  margin: 15px auto;
-`;
-
-const UrlWrapper = styled.div`
-  display: -webkit-box;
-  -webkit-line-clamp: 1;
-  -webkit-box-orient: vertical;
-  white-space: normal;
-  overflow: hidden;
-
-  color: #bdbdbd;
-  width: 100%;
-  height: 16px;
-  margin-bottom: 11px;
-`;
-
-const Input = styled.input`
-  padding: 12px 18px;
-  color: #bdbdbd;
-  border: 1px solid #e0e0e0;
-  border-radius: 5px;
-  font-size: 11px;
-  line-height: 16px;
-  width: 100%;
-  margin-bottom: 14px;
-`;
-
-const FolderSelector = styled.div`
-  font-weight: 500;
-  font-size: 14px;
-  line-height: 20px;
-  text-align: center;
-  color: #4285f4;
-  cursor: pointer;
-`;
-
-const StoreButton = styled(Button)<{ disabled: boolean }>`
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-  align-items: center;
-  padding: 9.5px 31px;
-  width: 88px;
-  height: 39px;
-  background: #e0e0e0;
-  border-radius: 56px;
-
-  ${({ disabled }) =>
-    disabled &&
-    css`
-      background-color: #4285f4;
-      cursor: pointer;
-    `}
-`;
-
-const ButtonWrapper = styled.div`
-  display: flex;
-  justify-content: flex-end;
-  margin-top: 65px;
-`;
+import { Icon } from "../components";
+import * as S from "./MainPage.style";
 
 const MainPage = () => {
   const [url, setUrl] = useState("");
@@ -109,8 +28,8 @@ const MainPage = () => {
   };
 
   return (
-    <Container>
-      <IconWraaper>
+    <S.Container>
+      <S.IconWraaper>
         <Icon name="logo" width={50} height={30} />
         <Icon
           name="btn_x"
@@ -118,8 +37,8 @@ const MainPage = () => {
           height={13}
           onClick={() => window.close()}
         />
-      </IconWraaper>
-      <ImageWrapper>
+      </S.IconWraaper>
+      <S.ImageWrapper>
         <img
           width={87}
           height={87}
@@ -127,14 +46,14 @@ const MainPage = () => {
           alt=""
           onError={handleErrorImg}
         />
-      </ImageWrapper>
-      <UrlWrapper>{url}</UrlWrapper>
-      <Input ref={inputRef} placeholder="북마크 이름" />
-      <FolderSelector>{folderSelector}</FolderSelector>
-      <ButtonWrapper>
-        <StoreButton disabled={buttonDisabled}>저장</StoreButton>
-      </ButtonWrapper>
-    </Container>
+      </S.ImageWrapper>
+      <S.UrlWrapper>{url}</S.UrlWrapper>
+      <S.Input ref={inputRef} placeholder="북마크 이름" />
+      <S.FolderSelector>{folderSelector}</S.FolderSelector>
+      <S.ButtonWrapper>
+        <S.StoreButton disabled={buttonDisabled}>저장</S.StoreButton>
+      </S.ButtonWrapper>
+    </S.Container>
   );
 };
