+++ web/pageComponents/folderlistComponents/components/SearchBar.style.ts
